services/api_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.21.1
pydantic==2.5.0
python-dotenv==1.0.0
